\name{myDirichlet}
\alias{myDirichlet}
\title{
Simulate from Dirichlet distribution
}
\description{
This function simulates random vectors from a Dirichlet distribution.
}
\usage{
myDirichlet(alpha)
}
\arguments{
  \item{alpha}{
Vector of positive numbers denoting the parameters of the Dirichlet distribution.
}
}
\author{
Panagiotis Papastamoulis
}

